refactor: Master, Viewer 컴포넌트 분리
